+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - github: Gists plugin and key maps
 - practicalli: conjure eval comment & hightlight options
 - community: neovide recipes for scaling functions & key mappings
+- termux: use local lsp language servers when mason install not supported
 
 ## Changed
 - readme: expand on description and use